requirements.txt
asgiref==3.8.1
boto3==1.37.9
botocore==1.37.9
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
dj3-cloudinary-storage==0.0.6
crispy-bootstrap5==0.7
Django==5.1.7
django-crispy-forms==2.3
django-embed-video==1.4.10
django-storages==1.14.5
django-cleanup==9.0.0
django-debug-toolbar==5.0.1
django-registration-redux==2.13
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing-extensions==4.12.2